clojure: Use lein-maven-s3-wagon for now.
s3-wagon-private is preferred because it is actively maintained and makes use
of ~/.aws credentials, but lein-maven-s3-wagon currently works on all active
projects.

Ideally, projects will migrate to include s3-wagon-private in the project.clj,
at which point it can become the S3 plugin of choice, and the env section of
repository auth can be removed from profiles.clj.